pyproject.toml
[project]
name = "aith-chatbot"
version = "0.1.0"
description = "Чат-бот для абитуриентов магистратур ИТМО (AI и AI Product)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.12.0",
    "html2text>=2020.1.16",
    "python-telegram-bot>=20.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "qdrant-client>=1.7.0",
    "langchain-text-splitters>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aith_chatbot"]
